Share sharded tensor CPU merge fallback
