+++ megatron/core/ssm/mamba_mixer.py
@@ -5,7 +5,6 @@
 # This source code is licensed under the Apache license found in the
 # LICENSE file in the root directory of this source tree.
 
-import gc
 import logging
 import math
 from dataclasses import dataclass, replace
@@ -35,6 +34,7 @@ from megatron.core.transformer.spec_utils import ModuleSpec, build_module
 from megatron.core.transformer.utils import (
     ensure_metadata_has_dp_cp_group,
     make_sharded_tensors_for_checkpoint,
+    sh_ten_merge_fn,
     sharded_state_dict_default,
 )
 from megatron.core.utils import (
@@ -1407,20 +1407,6 @@ def _split_tensor_factory(
             t.shape,
         )
         return chunk_sh_tens
-
-    @torch.no_grad()
-    def sh_ten_merge_fn(sub_state_dict):
-        try:
-            return torch.cat(sub_state_dict)
-        except (RuntimeError, torch.cuda.OutOfMemoryError) as e:
-            logger.warning(
-                f"CUDA OutOfMemoryError encountered during tensors merging."
-                f" Switching to CPU merge. (Error: {e})"
-            )
-            merged_sub_state_dict = torch.cat([t.cpu() for t in sub_state_dict])
-            gc.collect()
-            torch.cuda.empty_cache()
-            return merged_sub_state_dict
 
     return ShardedTensorFactory(
         orig_sh_ten.key, orig_sh_ten.data, sh_ten_build_fn, sh_ten_merge_fn, orig_sh_ten.replica_id
